refactor: 공통 dashboard 타입을 common.ts로 분리
IMetricsResponse, IRoasRanking을 common.ts 단일 출처로 이동
IPlatformPerformance는 IMetricsResponse를 extend해 중복 필드 제거
overview.ts, platform.ts는 common.ts에서 re-export
