+++ src/types/dashboard/common.ts
@@ -1,0 +1,21 @@
+// 공통 지표 응답 (overview/platform 공유)
+export interface IMetricsResponse {
+  clicks: number;
+  clickChangeRate: number;
+  impressions: number;
+  impressionChangeRate: number;
+  conversion: number;
+  cvrChangeRate: number;
+  ROAS: number;
+  ROASChangeRate: number;
+}
+
+// ROAS 순위 항목
+export interface IRoasRanking {
+  rank: number;
+  provider: string;
+  roas: number;
+  diffRate: number | null;
+  revenue: number;
+  adSpend: number;
+}
